refactor: add Subtitling namespace
