+++ src/LiveStreamingServerNet.StreamProcessor/Internal/Hls/Subtitling/Contracts/ISubtitleCueExtractor.cs
@@ -1,7 +1,7 @@
 ﻿using LiveStreamingServerNet.StreamProcessor.Internal.Hls.WebVtt;
 using LiveStreamingServerNet.StreamProcessor.Transcriptions;
 
-namespace LiveStreamingServerNet.StreamProcessor.Internal.Hls.Output.Contracts
+namespace LiveStreamingServerNet.StreamProcessor.Internal.Hls.Subtitling.Contracts
 {
     internal interface ISubtitleCueExtractor : IAsyncDisposable
     {
