+++ src/LiveStreamingServerNet.StreamProcessor/Internal/Hls/Subtitling/Contracts/ISubtitleTranscriber.cs
@@ -2,7 +2,7 @@
 using LiveStreamingServerNet.StreamProcessor.Transcriptions;
 using LiveStreamingServerNet.Utilities.Buffers.Contracts;
 
-namespace LiveStreamingServerNet.StreamProcessor.Internal.Hls.Output.Contracts
+namespace LiveStreamingServerNet.StreamProcessor.Internal.Hls.Subtitling.Contracts
 {
     internal interface ISubtitleTranscriber : IAsyncDisposable
     {
