+++ src/LiveStreamingServerNet.StreamProcessor/Internal/Hls/Subtitling/Contracts/ISubtitleTranscriberFactory.cs
@@ -1,7 +1,7 @@
 ﻿using LiveStreamingServerNet.StreamProcessor.Transcriptions;
 using LiveStreamingServerNet.StreamProcessor.Transcriptions.Contracts;
 
-namespace LiveStreamingServerNet.StreamProcessor.Internal.Hls.Output.Contracts
+namespace LiveStreamingServerNet.StreamProcessor.Internal.Hls.Subtitling.Contracts
 {
     internal interface ISubtitleTranscriberFactory
     {
