+++ src/LiveStreamingServerNet.StreamProcessor/Internal/Hls/Subtitling/SubtitleTranscriptionStreamFactory.cs
@@ -1,7 +1,7 @@
 ﻿using LiveStreamingServerNet.StreamProcessor.Transcriptions;
 using LiveStreamingServerNet.StreamProcessor.Transcriptions.Contracts;
 
-namespace LiveStreamingServerNet.StreamProcessor.Internal.Hls.Output
+namespace LiveStreamingServerNet.StreamProcessor.Internal.Hls.Subtitling
 {
     internal record SubtitleTranscriptionStreamFactory(
         SubtitleTrackOptions Options,
