+++ src/LiveStreamingServerNet.StreamProcessor/Internal/Utilities/FileHelper.cs
@@ -1,6 +1,6 @@
 ﻿using LiveStreamingServerNet.Utilities.Common;
 
-namespace LiveStreamingServerNet.StreamProcessor.Internal.Hls.Output.Writers.Utilities
+namespace LiveStreamingServerNet.StreamProcessor.Internal.Utilities
 {
     internal static class FileHelper
     {
